Refactor benchmarks to use Go benchmark framework
- Convert Test* benchmark functions to Benchmark* functions
- Move BenchmarkProduce from service/ to benchmarks/
- Split service/benchmark_test.go: TestSort stays in service/sort_test.go
- Replace custom benchmark runner with proper Go benchmarks:
  - BenchmarkProduceOperations, BenchmarkLeaseOperations
  - BenchmarkCompleteOperations, BenchmarkFullWorkflow, BenchmarkThroughput
- Update Makefile: make test excludes benchmarks, make benchmark runs go test -bench
- Change -logging flag to TEST_LOGGING env var for cross-package compatibility
- Update fuzz target path to ./service/...
